Add and resize images, update index.html
This will be the final commit for the project unless I decide to go back
and fix any lingering typos.

This commit adds the images directory to the main repository. It also
changes the size of the images on the recipe pages. I originally added
them in without any regard for sizing. Once I noticed that the recipe
for Blackened Chicken Tacos required viewers to scroll past a massive
image of a salad, I decided to fix the width of the images to 500px.
After saving and committing, I realized I should have fixed the HEIGHT
of the images, as that was the dimension causing issues.

I had also been committing from the /recipes directory and not main, so
I have been failing to upload and update the index.html file for the
website project. That has been rectified in this latest commit.